_include_key now works if used again on same Cs.
We were mutating the C if _prepend_key was used, which meant that if the
same list of Cs was used in more than one @parameters that used
_include_key=True, the key got prepended more than once.  The fix is
simple: don't mutate the C, return a new one.  That uses more memory,
but is not a cost worth worrying about.

Also dropped an unused import.